Organized auth, removed auth component and separated service and guard
