+++ src/app/_services/index.ts
@@ -1,3 +1,4 @@
+export * from './auth.service';
 export * from './google-analytics.service';
 export * from './loader-state.service';
 export * from './login-state.service';
