+++ src/app/app.module.ts
@@ -27,7 +27,6 @@ import {CIDService, NglComponent} from './ngl/ngl.component';
 import { IndicationsGraphComponent, GraphDataService } from './compound-data/indications-graph/indications-graph.component';
 // import { SearchResultTableComponent } from './compound-search/search-result-table/search-result-table.component';
 import {HttpClientModule, HttpClientXsrfModule, HTTP_INTERCEPTORS } from "@angular/common/http";
-import {AuthComponent, AuthGuard, AuthService} from './auth/auth.component';
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
 
 import {
@@ -97,8 +96,16 @@ import {
 } from './_components/index';
 
 import { ForgotPasswordComponent, LoginFailComponent, TermsComponent, UserRegistrationComponent } from './_dialogs/index';
+import { AuthGuard } from './_guards/auth.guard';
 import { LoaderInterceptorService } from './_interceptors/loader-interceptor.service';
-import { GoogleAnalyticsService, LoaderStateService, LoginStateService, SearchResultService, WDQService } from './_services/index';
+import {
+  AuthService,
+  GoogleAnalyticsService,
+  LoaderStateService,
+  LoginStateService,
+  SearchResultService,
+  WDQService
+} from './_services/index';
 import { ConfirmEmailComponent } from './confirm-email/confirm-email.component';
 import { ResetPasswordComponent } from './reset-password/reset-password.component';
 
@@ -172,7 +179,6 @@ export class MaterialModule {}
     CompoundSearchOptionsComponent,
     IndicationsGraphComponent,
     // SearchResultTableComponent,
-    AuthComponent,
     AuthGuard,
     EditItemComponent,
     DialogOverviewExample,
