--- src/app/auth/auth.component.css
@@ -1,3 +1,0 @@
-login {
-  align-content: flex-end;
-}
