--- src/app/auth/auth.component.html
@@ -1,1 +1,0 @@
-<auth-button></auth-button>
